interfaces: add implementation for dbus-name interface
The interface allows a snap to own a specified name on dbus system
bus. It is provied as implicit interface of the OS snap which
contains the dbus-daemon and therefor responsible to manage this
resource.

The interface will auto-connect and puts a simplified dbus policy
file in place which allows the root user to own the bus and to
send messages to the bus owner. No further security mediation
is put in place so by default no other snap is allowed to talk
to the service. This will be implemented with another interface.

Signed-off-by: Simon Fels <simon.fels@canonical.com>