interfaces: implement dbus-access interface
The dbus-access interface allows apps within the same snap to
communicate over dbus. It can be used in conjunction of the
dbus-name interface to setup up dbus service which don't need
to be accesible from another snap in the system.

Connecting the interface on between plugs and slots which are
not part of the same snap is prohibited and will cause the
interface connection to fail.

Signed-off-by: Simon Fels <simon.fels@canonical.com>