cmd/snap: do not allow classic snaps on other distributions
As of right now classic snaps are not supported on other distributions
than Ubuntu or Debian. This is because of those not allowing us to have
a global /snap directory, so its moved into /var/lib/snapd/snap instead.

The changed /snap directory breaks classic snaps as those have static
paths hardcoded to binaries/libraries in /snap for the sake of the
linker to work.